refactor(web): move the sidebar out of Layout into AppSidebar
Pure extraction, no behavior change: the nav column and its pieces move
to `AppSidebar`, which now owns the collapse and nav-group state it was
reading off the `useAppShell` bag. Layout keeps the header, the error
strip and the main region. A second caller for the sidebar arrives with
the mobile drawer.
